% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cva_tbl}
\alias{cva_tbl}
\title{Table of pre-computed critical values}
\format{
A data frame with 4 variables:
\describe{

\item{\code{B}}{B Bound on the square root of the average squared normalized
bias}

\item{\code{kappa}}{Bound on the kurtosis of the normalized bias}

\item{\code{cv}}{Critical value}

\item{\code{alpha}}{Determines confidence level, \eqn{1-\alpha}{1-alpha}.}

}
}
\source{
Computed using the \code{cva} function in this package
}
\usage{
cva_tbl
}
\description{
Table of pre-computed critical values
}
\keyword{datasets}
