% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{ebci}
\alias{ebci}
\title{EBCIs in applications}
\usage{
ebci(formula, data, se, weights, alpha = 0.1, kappa = NULL, tstat = FALSE)
}
\arguments{
\item{kappa}{Use pre-specified value for kurtosis (such as Inf). If NULL,
then compute it}
}
\description{
EBCIs in applications
}
