% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{w_eb}
\alias{w_eb}
\title{Empirical Bayes estimator and confidence intervals}
\usage{
w_eb(S, kappa = Inf, alpha = 0.05)
}
\arguments{
\item{S}{Signal-to-noise ratio \eqn{\mu_{2}/\sigma^2}{mu_2/sigma^2}, where
\eqn{\mu_{2}}{mu_2} is the second moment of
\eqn{\theta-X'\delta}{theta-X*delta}, and \eqn{\sigma^2}{sigma^2} is the
variance of the preliminary estimator.}

\item{kappa}{Kurtosis of \eqn{\theta-X'\delta}{theta-X*delta}.}

\item{alpha}{Determines confidence level, \eqn{1-\alpha}{1-alpha}.}
}
\value{
Returns a list with 3 components:
\describe{

\item{\code{w}}{Empirical Bayes shrinkage factor \eqn{w_{eb}}{w_eb}}

\item{\code{length}}{Normalized half-length of the corresponding confidence
interval, so that the interval obtains by taking the estimator based on
shrinkage given by \code{w}, and adding and subtracting \code{length} times
the standard error \eqn{\sigma}{sigma} of the preliminary estimator.}

\item{\code{m2}}{Second moment of the normalized bias, \code{1/m2}.}

}
}
\description{
Compute empirical Bayes shrinkage factor \eqn{w_{eb}}{w_eb} and the
normalized half-length of of the associated robust Empirical Bayes confidence
interval (EBCI).
}
\examples{
w_opt(1, 3)
## No constraint on kurtosis yields doesn't affect shrinkage, but yields
## larger half-length
w_opt(1, Inf)
}
\references{
{

\cite{Armstrong, Timothy B., Kolesár, Michal, and Plagborg-Møller, Mikkel
(2020): Robust Empirical Bayes Confidence Intervals,
\url{https://arxiv.org/abs/2004.03448}}

}
}
