% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{wl2}
\alias{wl2}
\title{Optimal shrinkage for Empirical Bayes estimator}
\usage{
wl2(S, kappa = Inf, alpha = 0.05)
}
\arguments{
\item{S}{average signal, estimate of standard deviation of
\eqn{\theta}{theta}}

\item{kappa}{kurtosis of \eqn{\theta}{theta}}

\item{alpha}{determines CI level}
}
\description{
Optimal shrinkage for Empirical Bayes estimator
}
