% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{wopt}
\alias{wopt}
\title{Optimal shrinkage for Empirical Bayes confidence intervals}
\usage{
wopt(S, kappa, alpha = 0.05, cv_tbl = NULL, maxbias = 10)
}
\arguments{
\item{S}{average signal, estimate of standard deviation of
\eqn{\theta}{theta}}

\item{kappa}{kurtosis of \eqn{\theta}{theta}}

\item{alpha}{determines CI level}

\item{cv_tbl}{optionally, supply a data frame of critical values.}

\item{maxbias}{maximum bias of estimator, determines an upper bound on the
amount of shrinkage}
}
\value{
TODO
}
\description{
Optimal shrinkage for Empirical Bayes confidence intervals
}
\examples{
wopt(1, 3, cv_tbl=cva_tbl[cva_tbl$kurt==3 & cva_tbl$alpha==0.05, ])
}
