% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cva}
\alias{cva}
\title{Compute average coverage critical value under moment constraints.}
\usage{
cva(B, kappa = Inf, alpha = 0.05, check = TRUE)
}
\arguments{
\item{B}{Bound on the square root of the average squared normalized bias,
\eqn{\sqrt{m_{2}}}{sqrt(m_2)}}

\item{kappa}{Bound on the kurtosis of the normalized bias, \eqn{\kappa}{kappa}}

\item{alpha}{Determines confidence level, \eqn{1-\alpha}{1-alpha}.}

\item{check}{If \code{TRUE}, verify accuracy of the solution by checking that
the implied least favorable distribution satisfies the \code{B} and
\code{kappa} constraints and yields the same non-coverage rate. If this
fails (perhaps due to numerical accuracy issues), solve a finite-grid
approximation (by discretizing the support of the bias) to the primal
linear programing problem, and check that it agrees with the dual
solution.}
}
\value{
Returns a list with 4 components: \describe{

\item{\code{cv}}{Critical value for constructing two-sided confidence intervals.}

\item{\code{size}}{The argument \code{alpha}.}

\item{\code{x}}{Support points for the least favorable distribution for the squared
normalized bias, \eqn{b^2}.}

\item{\code{p}}{Probabilities associated with the support points.}

}
}
\description{
Computes the critical value \eqn{cva_{\alpha}(m_{2}, \kappa)}{cva_alpha(m_2,
kappa)} from Armstrong, Kolesár, and Plagborg-Møller (2020).
}
\examples{
## Critical value without imposing a constraint on kurtosis
cva(1, kappa=Inf)
## With a constraint
cva(1, kappa=3)
}
