% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cva}
\alias{cva}
\title{Compute average coverage critical value under moment constraints.}
\usage{
cva(B, kurt = Inf, alpha = 0.05, check = TRUE)
}
\arguments{
\item{B}{Bound on the square root of the average squared bias}

\item{kurt}{Bound on the kurtosis of the bias}

\item{alpha}{Determines CI level, \eqn{1-\alpha}{1-alpha}.}

\item{check}{Check critical value by solving linear program}
}
\value{
Critical value for constructing two-sided confidence intervals.
}
\description{
Compute average coverage critical value under moment constraints.
}
\examples{
## Without imposing a constraint on kurtosis
cva(1, kurt=Inf)
## With a constraint
cva(1, kurt=3)
}
