% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cva_tbl}
\alias{cva_tbl}
\title{Table of pre-computed critical values}
\format{
A data frame with 18 rows and 19 variables:
\describe{
\item{B}{B Bound on the square root of the average squared bias}
\item{kappa}{Bound on the kurtosis of the bias}
\item{cv}{Critical value}
\item{alpha}{Determines CI level, \eqn{1-\alpha}{1-alpha}.}
}
}
\source{
Computed using the \code{cva} function in this package
}
\usage{
cva_tbl
}
\description{
Table of pre-computed critical values
}
\keyword{datasets}
