% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{w_eb}
\alias{w_eb}
\title{Empirical Bayes estimator and confidence intervals}
\usage{
w_eb(S, kappa = Inf, alpha = 0.05)
}
\arguments{
\item{S}{Square root of the signal-to-noise ratio
\eqn{\sqrt{\mu_{2}}/\sigma}{sqrt(mu_2)/sigma}, where \sqrt{\mu_{2}}{mu_2}
is the variance of \eqn{\theta}{theta}}

\item{kappa}{Kurtosis of \eqn{\theta}{theta}}

\item{alpha}{determines CI level}
}
\value{
(TODO)
}
\description{
Compute linear shrinkage factor \eqn{w_{eb}}{w_eb} and normalize length of of
the associated robust Empirical Bayes confidence interval (EBCI).
}
