% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{w_eb}
\alias{w_eb}
\title{Empirical Bayes estimator and confidence intervals}
\usage{
w_eb(S, kappa = Inf, alpha = 0.05)
}
\arguments{
\item{S}{Square root of the signal-to-noise ratio
\eqn{\sqrt{\mu_{2}}/\sigma}{sqrt(mu_2)/sigma}, where
\eqn{\sqrt{\mu_{2}}}{mu_2} is the variance of
\eqn{\theta-X'\delta}{theta-X*delta}, and \eqn{\sigma^2}{sigma^2} is the
variance of the preliminary estimator.}

\item{kappa}{Kurtosis of \eqn{\theta-X'\delta}{theta-X*delta}.}

\item{alpha}{Determines confidence level, \eqn{1-\alpha}{1-alpha}.}
}
\value{
Returns a list with 3 components:
\describe{

\item{\code{w}}{Empirical Bayes shrinkage factor \eqn{w_{eb}}{w_eb}}

\item{\code{length}}{Normalized half-length of the corresponding confidence
interval, so that the interval obtains by taking the estimator based on
shrinkage given by \code{w}, and adding and subtracting \code{length} times
the standard error \eqn{\sigma}{sigma} of the preliminary estimator.}

\item{\code{B}}{Square root of the normalized bias, \eqn{1/S}.}

}
}
\description{
Compute empirical Bayes shrinkage factor \eqn{w_{eb}}{w_eb} and the
normalized half-length of of the associated robust Empirical Bayes confidence
interval (EBCI).
}
\examples{
w_opt(1, 3)
## No constraint on kurtosis yields the same shrinkage, but larger half-length
w_opt(1, Inf)
}
