% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb.R
\name{w_opt}
\alias{w_opt}
\title{Optimal shrinkage for Empirical Bayes confidence intervals}
\usage{
w_opt(S, kappa, alpha = 0.05, cv_tbl = NULL)
}
\arguments{
\item{S}{Signal-to-noise ratio \eqn{\mu_{2}/\sigma^2}{mu_2/sigma^2}, where
\eqn{\mu_{2}}{mu_2} is the second moment of
\eqn{\theta-X'\delta}{theta-X*delta}, and \eqn{\sigma^2}{sigma^2} is the
variance of the preliminary estimator.}

\item{kappa}{Kurtosis of \eqn{\theta-X'\delta}{theta-X*delta}.}

\item{alpha}{Determines confidence level, \eqn{1-\alpha}{1-alpha}.}

\item{cv_tbl}{Optionally, supply a data frame of critical values.
\code{cva(m2, kappa, alpha)}, for different values of \code{m2}, such as
a subset of the data frame \code{\link{cva_tbl}}, that matches the
supplied values of \code{alpha} and \code{kappa}. The data frame needs to
contain two variables, \code{m2}, corresponding to the value of second
moment of the normalized bias, and \code{cv}, with the corresponding
value of \code{cva(m2, kappa, alpha)}. If non \code{NULL}, for the
purposes of optimizing the shrinkage factor, compute the critical value
\code{cva} by interpolating between the critical values in this data
frame, instead of computing them from scratch. This can speed up the
calculations.}
}
\value{
Returns a list with 3 components: \describe{

\item{\code{w}}{Optimal shrinkage factor \eqn{w_{opt}}{w_opt}}

\item{\code{length}}{Normalized half-length of the corresponding confidence
interval, so that the interval obtains by taking the estimator based on
shrinkage given by \code{w}, and adding and subtracting \code{length} times
the standard error \eqn{\sigma}{sigma} of the preliminary estimator.}

\item{\code{m2}}{The second moment of the normalized bias,
\eqn{(1/w-1)^{2}S}{(1/w-1)^2*S}}}
}
\description{
Compute linear shrinkage factor \eqn{w_{opt}}{w_opt} to minimize the length
of the resulting Empirical Bayes confidence interval (EBCI).
}
\examples{
w_opt(1, 3)
## Use precomputed critical value table
w_opt(1, 3, cv_tbl=cva_tbl[cva_tbl$kappa==3 & cva_tbl$alpha==0.05, ])
}
\references{
{

\cite{Armstrong, Timothy B., Kolesár, Michal, and Plagborg-Møller, Mikkel
(2020): Robust Empirical Bayes Confidence Intervals,
\url{https://arxiv.org/abs/2004.03448}}

}
}
